OK, there was some crazy ajax weirdness with express.bodyParser, but it seems to be resolved now.
Got log in working with mongo and starting to hook up backbone to it.

GET for /entries seems to work and POST to /entries also.

BUG currently is the view is getting rendered twice and thus 2 copies are visible since we prepend the view every time.

Committing now because it took 2 days to sort this stuff out.